// projects/keyboard/src/lib/enums/keyboard-key.enum.ts
/**
 *
 *
 * @export
 * @enum {number}
 */
export enum KeyboardKey {
    Alt = 'Alt',
    AltGraph = 'AltGraph',
    AltLk = 'AltLk',
    CapsLock = 'CapsLock',
    Enter = 'Enter',
    Shift = 'Shift',
    Space = 'Space',
    Tab = 'Tab',
    Backspace = 'Backspace',
    ArrowRight = 'ArrowRight',
    ArrowLeft = 'ArrowLeft'
}